Move the models for dogapi endpoints to Endpoints/Breeds folder to follow VSA
