+++ src/DogApi/Endpoints/Breeds/Models/RandomBreedImageByBreedRequest.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 
-namespace DogApi.Models.Breeds
+namespace DogApi.Endpoints.Breeds.Models
 {
     public class RandomBreedImageByBreedRequest
     {
