+++ src/DogApi/Endpoints/Breeds/Models/RandomBreedImageByBreedResponse.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using System.Text.Json.Serialization;
 
-namespace DogApi.Models.Breeds
+namespace DogApi.Endpoints.Breeds.Models
 {
     public class RandomBreedImageByBreedResponse
     {
